+++ restoran/v1_urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from restoran.v1_views import forum_by_restoran
+from forum.v1_views import forum_by_restoran
 from minuman.v1_views import minuman_by_restoran
 
 app_name = "v1_restoran"
